+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixed handling of extras.
 - Fixed duplicate entries in both sdist and wheel.
 - Fixed excluded files appearing in the `package_data` of the generated `setup.py`.
+- Fixed transitive directory dependencies installation.
 
 
 ## [0.11.5] - 2018-09-04